+++ main-app/src/DatabaseServiceProvider.ts
@@ -1,7 +1,32 @@
+import { Middleware } from "neko-router/src";
+import { Request, Response } from "neko-router/src/types";
 import { PostgresqlConnection } from "neko-sql/src/databases/postgresql/PostgresqlConnection";
-import { Connection, PoolClient, PoolConfig } from "pg";
+import { PoolConfig } from "pg";
+import { Connection } from "neko-sql/src/Connection";
+import { BaseModel } from "neko-orm/src/baseModel";
+import { ctx } from "neko-http/src";
 
-export class DatabaseServiceProvider {
+export class DatabaseServiceProvider extends Middleware {
+  async call(
+    req: Request,
+    res: Response,
+    next: () => Promise<void>,
+  ): Promise<void> {
+    const db = DatabaseServiceProvider.getInstance();
+    const conn = await db.getConnection();
+    // @ts-ignore
+    ctx().get<Request>("request").context = { dd: "cc" };
+    try {
+      ctx().set("db", conn);
+      BaseModel.setConnection(() => ctx().getOrThrow<Connection>("db"));
+      await next();
+    } catch (err) {
+      throw err;
+    } finally {
+      await conn.disconnect();
+    }
+  }
+
   private static instance: DatabaseServiceProvider;
   private conn: PostgresqlConnection | undefined;
 
@@ -21,7 +46,6 @@ export class DatabaseServiceProvider {
       user: process.env.DB_USER,
       password: process.env.DB_PASSWORD,
       port: parseInt(process.env.DB_PORT || "5432"),
-      max: 3,
     };
 
     const conn = new PostgresqlConnection(db_config);
