+++ main-app/src/router.ts
@@ -1,0 +1,13 @@
+import { Router } from "neko-router/src";
+
+function createSingleton<T>(func: (...args: any[]) => T): () => T {
+  let instance: T;
+  return (...args: any[]): T => {
+    if (!instance) {
+      instance = func(...args);
+    }
+    return instance;
+  };
+}
+
+export const router = createSingleton(() => new Router())();
