+++ main-app/src/test_server.ts
@@ -1,4 +1,3 @@
-import { Router } from "neko-router/src";
 import { HttpServer } from "neko-http/src";
 import { Request, Response } from "neko-router/src/types";
 import { HttpError } from "http-errors";
@@ -6,7 +5,9 @@ import { wait } from "neko-helper/src/time";
 import { DatabaseServiceProvider } from "./DatabaseServiceProvider";
 import { ctx } from "neko-http/src";
 import { Connection } from "neko-sql/src/Connection";
-import { BaseModel } from "neko-orm/src/baseModel";
+import { router } from "./router";
+
+import "./routes";
 
 let server = new HttpServer();
 
@@ -23,26 +24,8 @@ server.setErrorHandler(async (err: Error, req: any, res: any) => {
   res.end(JSON.stringify({ error: "Internal Server Error" }));
 });
 
-let router = new Router();
-
 // load database connection
-router.addGlobalMiddleware(
-  async (req: Request, res: Response, next: () => Promise<void>) => {
-    const db = DatabaseServiceProvider.getInstance();
-    const conn = await db.getConnection();
-    // @ts-ignore
-    ctx().get<Request>("request").context = { dd: "cc" };
-    try {
-      ctx().set("db", conn);
-      BaseModel.setConnection(() => ctx().getOrThrow<Connection>("db"));
-      await next();
-    } catch (err) {
-      throw err;
-    } finally {
-      await conn.disconnect();
-    }
-  },
-);
+router.addGlobalMiddleware(DatabaseServiceProvider);
 
 router.addGlobalMiddleware(
   async (req: Request, res: Response, next: () => Promise<void>) => {
@@ -67,20 +50,10 @@ router.addRoute(
   },
 );
 
-function InjectedValue(value: any) {
-  return function (
-    target: any,
-    propertyKey: string | symbol,
-    parameterIndex: number,
-  ) {
-    return 1;
-  };
-}
-
 /*
 CREATE TABLE cats (
   id SERIAL PRIMARY KEY,
-  name VARCHAR(255) UNIQUE
+  name VARCHAR(255) UNIQUE NOT NULL,
 );
 */
 router.addRoute(
@@ -100,6 +73,7 @@ router.addRoute(
         sql: "insert into cats (name) values ($1)",
         bindings: [req?.query?.name as string],
       });
+      console.log("inserted", req?.query?.name);
       await db.commit();
       error = "success";
     } catch (err) {
